requirements.txt
# Core
Django==5.2
djangorestframework==3.15.2

# Utilities
django-filter==24.3
django-cors-headers==4.6.0
pillow==11.0.0

# Database
psycopg[binary]==3.2.3

# Development
django-debug-toolbar==4.4.6
django-extensions==3.2.3

# Background tasks (optional, for later)
celery==5.4.0
redis==5.1.0


# API authtication and documentation
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0